Try out converting Model (partly) to a union type
The model now contains a "mode" which varies depending on the user login. The
data in the model can thus be restricted to whatever makes sense for the
current user, rather than keeping all possible states in a flat record type.

For example, Login.Model now only exists if the user hasn't already logged in
(the app is in Anon state).

Another possibility is moving the answers form to the student mode. Others will
probably emerge as the app gets more complicated.