Prepare Elm code generator for use of AuthProtect
Using AuthProtect to implement custom authentication for an API won't work with
servant-elm (see https://github.com/mattjbray/servant-elm/issues/11).

This change is simlilar to the one suggested in that issue, but rather than
doing nothing, it adds an Authorization header to the request (which results
in another parameter being passed into the generated API request functions
to set the header value). Elm doesn't provide a way to modify the request data
once it has been created so without this it would be impossible to set a
header.