Extract sortable table customizations for reuse
Create a Bootstrap module since the customizations are mostly about adding
Bootstrap classes (apart from the fix to add the missing tr element to the
table header). More Bootstrap stuff will probably end up in here to make
creating rows, columns and so on easier.

Also extracts the table config into a separate function since it is static
and I don't think the Elm compiler is smart enough to move it to the top
level automatically.