Implement creating new student accounts.
A list of names is passed via the API and the response returns the Student
objects along with the username and password.

New accounts are listed above the overall list of students to allow the
teacher to make a not of the account names and passwords. A print button
and a means of deleting this list should also be added later.