Get Hasql working for Story data type
Works for storing and retrieving all the story data.

Dealing with the composite list of dictionary entries is a bit tricky when
doing an insert/update and requires a workaround where the entry is unzipped
into arrays and then recombined in the SQL and cast to the dict_entry type.
See https://github.com/nikita-volkov/hasql/issues/65 for more information.