Fix Answer row decoder and add optional story parameter
The answers query was failing because it was attempting to decode the storyId
column as a UUID rather than plain text.

It's not possible to add a query parameter for the storyId to retrieve answers
for a specific story. In this case the student's own ID will be left out of the
query and all answers for that school/story combination will be selected.