Post-processing of story content to add word lookups
This is a first attempt at adding dictionary lookups to the story view.

This cannot be done in Elm, since there is no way of customizing the markdown
processing beyond adding markdown syntax. Even if we allow non-strict HTML
content and pre-process the markdown content to add modifying elements to the
highlighted words within the story, it won't be possible to add event listeners
to the elements due to the CSP disallowing inline Javascript.

Instead a port is used to invoke a Javascript function when the story page is
shown. The function uses requestAnimationFrame to wait until the storycontent
element is present, then adds child elements and click listeners to highlight
the words.

Currently the words are just matched as the <em> content within the story (which
are inserted in the Elm code before invoking Markdown.toHtml. The event listener
re-invokes another port which passes a DictLookup message to Elm. This doesn't
do anything yet, but the basic flow of passing content to Javascript and
receiving events as the user clicks on words works OK.

- The story words (and meaning indices) need to be used to make sure other em
elements aren't modified, and that the correct meaning is returned.
- It might be preferable to do all the processing in JS, rather than adding em
elements in Elm at all (which is a bit of a hack).