Start of major rework based on the elm-spa-example
The application was seriously in need of a substantial refactoring and Richard
Feldman recently released this example app, which seems like a good model for
restructuring the 3ML Elm code. It includes best practices gleaned from
real-world experience. Following a standard structure will make the code
easier to understand for others.

So far the non-teacher functionality has been re-implemented. The
elm-spa-example code loads data on a per-page basis, and using per-page models
allows us to get rid of a lot of state that was previoulsy cluttering up the
top level models. However, we still want to be able to cache some data (such as
the stories) between different pages, so the role of Data.Session is expanded
to include caching of data.