Add answers to the story page model
This will allow the answers to be shown once they form has been submitted. It
also allows a check to be made on whether the student has already completed the
story, which will allow the answers to be shown immediately (without creating
the form).

Changes include:

* Loading the answers when the page is loaded (but only if the user is logged in
and the story exists).
* Subject ID is now retained in the client user data, to allow its use when
checking if the user has already completed the story questions.
* Adding an extra query parameter to the answers API to optionally supply
the student Id. This will allow a teacher to query for answers or a specific
story or student. The DB code doesn't support the use of both at the same time
yet - the student ID will take priority.
* Fix SQL error where story ID was being case to uuid.