Fix "too much recursion" bug in dict lookup
The problem occurs when there is a circular loop between dictionary entries.
Instead rendering all the definitions recursively as a tree, we acumulate them
in a dictionary, skipping any subdefinition if the dictionary already contains
it. The definitions are finally converted into a list and the entries converted
to Html.

Fixes #10, #8.