Make sure clicked-on word appears first in definitions
The downside of using a Dict to accumulate the word definitions and
sub-definitions is that we lose the ordering. This change makes sure that the
selected word from the story appears first in the list.

Ideally we should have the interface only display sub-defintions incrementally
when the user asks for them, particularly when there is a cascade of further
definitions (see, for example, the word "genome" in story 285).