Use school_id in class member update
The school_id is obtained from the user's access token and they should only be
able to modify data from that school. This change uses overlapping foreign keys
in the student_class table to ensure that the class and student entities both
have the same school and also that it matches the one used in the query. This
requires adding a redundant school_id column to the join table, and
corresponding unique constraints to the student and class tables, which match
the foreign key constraints.