Add 'hidden' and 'deleted' columns to student, and 'locked' to login
The student deleted flag is a timestamp so that we can just mark the account
for future deletion in the case where it has been used.

Login also needs a flag to prevent deleted accounts from logging in. In this
case though, it may be reused for other purposes (e.g. locked out by an admin),
so the flag is called 'locked'.