Store use information in browser session storage
This allows the app to be reloaded without having to log in again. We use
session storage rather than local storage so that closing the browser will be
equivalent to logging out.

A port is invoked when a user logs in to store the user data and the data is
passed to the Elm application as flags on initialization. Most of the changes
are implementing the encoders and decoders for the frontend user data
structure.

Resolves #35.