Separate page loaded messages from page messages
These are essentially two separate things.

The updatePage function was also getting rather large and the wildcard match
means we don't get any warnings when we forget to handle a message. Creating a
PageLoaded message means this can be matched separately and we get back the
compiler guarantees for these messages at least. It also removes the need to
match on the "page" submodel for these messages.

Also renamed MsgNew back to Msg, which was left over from the original
conversion to the elm-spa-example style.