Detect 401 response during page loads
The client now detects a 401 response, invalidates the current session and asks
the user to log back in.

The implementation is a bit hacky since the logic is in the Errored page module,
which now needs the Http error passed to the pageErrored function. This in turn
means it can't be called for non-Http errors, but it's easy to just use the
PageLoadError constructor with a message directly in those cases.

Closes #36.