Introduce SchoolAdmin role
It's been decided that certain admin functionality should only be available to
some teachers, so we need to differentiate between teacher accounts. This adds
an additional SchoolAdmin user type at the DB enum level and above, with
the teacher role being flagged as an admin or not.

Only admins can

- create new student accounts (the main motivation is to avoid duplicates)
- delete accounts
- hide accounts

All teachers can currently change the level of a student or change their
username or password (if they forget it, for example). These permissions may be
reviewed later.

See #40.