Basic story editor support
The editor page shows the markdown in a textarea and a story preview beside it.
See #23.

Had to remove the date from the Story type because of incompatibilties between
the encoded Elm date type and decoding it into a UTCTime on the backend. Going
the other way (to the client) is OK, but parsing fails on the submitted story
data when updating. It's not really clear what elm-export should do
(see elm-export/issues/38).