Implement leader board
Adds a "leaderboard" view to the datbase which contains the scores calculated
from the story totals, as well as the overall position in the table (calculated
using row_num). This means even though it will only be queried by school,
students will be able to see where they are compared with users outside their
school too.

Since the data is just a materialized view, it can be dropped and replaced
with a different version without losing any data, so the scoring can be
implemented directly in the database.