Sort stories to show newest ones first
The sorting is done in the client.

Also converted the 'id' column in the story to an integer which makes sorting
easier and will make new stories use the same sequence (the column is a
postgresql 'serial' type).

Note that to avoid conflicts, when setting up a new DB, the value of the column
sequence should be set to the maximum existing value. For example:

select setval('story_id_seq', 922);

Closes #56.