Explicitly roll back SQL transaction on error
Postgresql doesn't automatically roll back if there's an error and we have
manually started a transaction. So we get the error:

"25P02" "current transaction is aborted, commands ignored until end of
transaction block"

when we try to use the connection again.

This explicitly wraps the session with a rollback command when there's an
error.