Make student accounts enabled on creation
The DB default is to set the active column to 'false', which causes a problem
when logging in as a student (and there is no mechanism for a teacher to make
an account active). To fix this we explicitly set the 'active' flag in the SQL
for inserting an account.

Fixes #58.