Refactor story view
The image width updating stuff is now part of the view state, rather than being
duplicated in the editor and story pages which used the view. Adding it to the
story browser for the "find story" page will be yet another use so it's worth
the extra nesting of models.