Add forward and previous navigation through stories
Both the story preview in FindStory and the editor now support moving back
and through a zipper of stories. Switched to an infinite zipper rather than the
standard list zipper so that we just wrap round at either end.

Also added use of the window size when deciding how to render the story image.