Implement lazy-loading of tiles in FindStory
When displaying the tiles view, new tiles are only loaded when scrolled to the
bottom of the page.

- On scroll events we check whether we are at the bottom of the page (the app
div) and increase the number of visible tiles if we are.
- A small margin is added to detect that the bottom of the app div is visible
(without this the bottom is a small rounding value different from the height of
the window).
- The number of tiles to show is estimated from the window size.
- After loading more tiles, we check if the last tile is visible in case the
estimate was wrong. If it is, we load more until it's no longer necessary.
- Make sure we don't check for visibility if we are at the end of the story
list or we'll end up stuck in a loop trying to show more content.

See #46.