Add "enabled" and "archived" flags to story
If a story is disabled it may be needing corrections or be in development. An
archived one is just flagged in the database and excluded from all queries.

An editor can see the story status and filter the story list to only show
disabled stories. At some point the editor will be updated to allow the status
to be changed.

Closes #64.