Fix teacher account activation
Don't how activation button after account is activated.

The servant API was also using PostNoContent for the activation endpoint which
is wrong. It results in a NoContent response, but the API was attempting to
return the ID of the activated account. As a result, Elm would never get the
activation response message.