Clear selected stories when creating anthology
It's unlear to a user that the anthology has been created if the list of
selected stories remains at the top of the screen. The client now empties
the selected stories list and switches to the anthologies view after creating
an anthology successfully, so that the anthology is visible at the top of the
screen.