Updates to anthology view and creation
- Stories are now collected in a "basket" analogous to online shopping.
- Small story tiles are used to render the basket and the anthology contents.

The basket is shown even when empty. If empty a message is shown saying the
basket is empty and giving the teacher a hint that it is possible to collect
stories in it.

Using small tiles makes the basket take up much less space than a table.
Similarly for anthologies.