Implement alerts
Allows pages to send alerts in a standard fashion, using the session, for
example, to report the success (or failure) of an operation.

Alerts are displayed using bootstrap alert style and can be dismissed using the
close icon. This requires a new global message constructor - CloseAlert.

Alerts which have not been closed will close automatically after a short delay
using a timer subscription.

Resolves #84.