Switch to using UUIDs in DB types
This is more robust than just using text and avoids unnecessary casting in SQL.

Originally I thought there might be problems with deriving Elm types for UUIDs
but it's actually trivial to use UUIDs on the back end and text in the Elm
types. It's probably also possible to use UUIDs in Elm, but there's little need
for that yet.