Fix issue with preventDefaultOn not working for href links
When using Browser.application in Elm 0.19, it is not possible to prevent a
click from sending the URL change request. Previously links with preventDefault
set would not change the URL in the browser. Now, regardless of whether this is
set, the onUrlRequest handler for the application is invoked.

The workaround here is to ignore any URL changes with an empty fragment when
handling LinkClicked messages (the href is set to "#" for the links in question).

Consequently the onClickPreventDefault function is no longer required and has
been removed.

The onClickStopPropagation function has also been removed as it wasn't being
used anyway.