Implement infinite zipper for story list
This was previously a separate package which isn't available for Elm 0.19.
However it's easy to implement the same functionality in a simple wrapper
round the 'listzipper' package.

So List.InfiniteZipper is replaced by List.Zipper.Infinite and uses the same
Zipper type as List.Zipper. The only functions we use which are different are
'next' and 'previous', returning 'a' rather than 'Maybe a' for the infinite case.