Add tighter DB constraint on story clarify_word column
Make sure it is contained in either the content or the title. This also detected
a few rows with bad data which were fixed and prevents the calrify word from
getting out of sync with the content in future.

This would probably be better implemented as a DB function which would be
tidier and avoid so many calls to 'lower'.

The frontend editor also needs to be fixed since it will break if the changed is
saved without the clarify word being present.