Support for 2FA via TOTP
If a hex-encoded TOTP key is added to the database for a user, the system will
prompt them for the code during login and check it as part of the subsequent
authentication request. A custom response code (462) is used to inform the
client that the account needs to use 2FA.

Intended for use with admin accounts at some point but works for any if the
key is set.