Add basic admin dashboard
This is a separate app which either shows a login or the dashboard.
A logged in Admin user will see the data from the admin/stats endpoint
rendered in the page.

Uses elm-visualization for now but this is kind of low-leve and quite a lot of
work so I may take a look at elm-vegalite instead and just server up the
appropriate JSON from the backend.