Story editor improvements
Support editing of more fields:

- Clarify word
- Title
- Level
- Curriculum
- Qualification
- Enabled status

Also detects when something has been change by comparing the original copy of
the story in the zipper with the modified one and enables/disables the save
button accordingly.

Validates the text fields and makes sure that the clarify word is contained in
either the content or the title (matching the constraints on the DB column
introduced in 26eaec934899).

Closes #63.