Add school information to the dashboard
Pulls back a list of schools, their teachers and the number of
registered students for eacher school.

Since the dashboard is pure JSON, this only involves changes in
Postgres and the admin client. A new DB function school_data()
is used to collate the information.

UI is very basic for now but will be styled later.