Change admin graph dashboard to use vega-lite
This allows us to produce nicely-formatted graphs with a minimal amount of code,
compared to the lower-level elm-visualization package. The rendering takes place
in Javascript though so it's hard to know what's going on.

The postgres function for retrieving story activity has also been changed to
provide more tabular data direct to the frontend as this is easier to handle in
a vega-lite specification.

Using vega can cause some problems with CSP. It requires 'unsave-eval'
for the script-src directive and 'unsafe-inline' for style-src. I have these
enabled for the admin app page in nginx but not for the main app.

The school data is now also formatted into cards using tailwinds CSS. This
clashes with the login page's bootstrap setup but it still works. I'll probably
move off bootstrap completely at some point in favour of a more lightweight
CSS setup.