Fix story printing
We want to use a smaller font to fit the story on one page if possible. Change
story view to accept a 'Maybe Settings', otherwise the 'style' attributes
created from the settings override the small print font. When a teacher is
viewing the story page, we pass Nothing for the settings so the story view.