Initial code for "work queue" feature
Allows a student to add an anthology to their "work queue" which appears on the
home page, meaning they don't have to re-visit the "find a story" page each time
they complete a story. They can collect stories they want to work on and the
queue will be saved between logins.

The story queue is persisted as part of the user settings, which makes the code
a bit hacky but means no additional backend code or database schema changes are
needed.