Generate passwords using a separate word list table
The original commit c8d15138863 used the words from the science dictionary as a
password generation list. There have been complaints that this is too
complicated for some children using the system.

We now have a separate word_list database table which can be used to generate
passwords.

As before, a list of words is selected from the database table and the words are
concatenated until the specified minimum length is reached. The words now have
'-' characters interspersed between them to make it easier to see the boundaries.

Note that using 'SELECT word FROM word_list ORDER BY random()' requires a full
scan of the table followed by a random sort. A much faster option is to use
'SELECT word FROM word_list TABLESAMPLE BERNOULLI (1)' which samples 1% of the
table rows. However the time for both is still small enough that it's not too
much of a concern at the moment (~10ms vs 1ms on my laptop).