Make Session an opaque type
The use of the global Session instance has become kind of pervasive which
increases complexity as more things are added to it (since everything has
access). This is a first step on limiting access to the internals to allow
future refactoring.

The Cache has been split out into a separate module with a view to changing
some functions which currently accept a Session parameter just to load cache
data to use this narrower type instead. Eventually the two can possibly be
separated allowing the session to be limited to containing user login data and
so on.