Re-sort stories in FindStory based on table order
I didn't realise before that elm-sortable-table exposes a getSortedData function
which sorts a list based on the current table state. It makes sense to reorder
the backing list of stories every time the table sort order changes. This way
the story browser opens with the same story ordering as the user has chosen via
the table.