Honour useSmallTiles value in StoryTiles view
The width, height and flexbox CSS are chosen differently depending on the
whether useSmallTiles is set or not, so that small tiles display better. This is
currently only used in the anthologies view.

Also changed the resizing behaviour of tiles depending on the viewport size.
For Tailwings "lg" display and above, the tile will no longer resize which
should avoid over-wide tiles that occur when only one or two are on a row.