Add "recent stories" filtering to FindStory page
This adds a select to the story filter which allows showing only stories
created in the last month, 3 months, 6 months or past year.

The creation date is represented by an epoch seconds value in the story and the
periods are approximate (simple multiples of the number of seconds in a day
rather than actual months).