Upgrade to Tailwind 1.0.1
Mostly following the upgrade guide from the Tailwind website for new colour
names and other changes.

Also made some styling changes in the process:

Most pages are now designed to fit the "container" width (i.e. the width of the
current "breakpoint"). The "xl" breakpoint is omitted. A Tailwinds Elm module was
added to expose the breakpoint sizes (which also changed in 1.0) and make it
more obvious how some of the magic numbers previously used were calculated.

The StoryTiles view has been rewritten to use the approach described in "grids"
on the Tailwind site. A fixed number of tiles are displayed per row, depending
on the breakpoint, starting at 2 and going up to 4. A simple wrapping flex
container with proportional widths on the elments is all that's required to
achieve this (e.g. w-1/4 for 4 tiles per row) and it simplifies a lot of the
other code in the module.