Fix dictionary words bug in story view
When viewing a story, the Javascript which post-processes the story looks for
the "storycontent" div to decide what to run on. If it is null (hasn't been
rendered yet), it calls requestAnimationFrame to wait.

However if we are moving directly from one story page to another (e.g. by
altering the story id number directly in the URL) then there is already a
storycontent div present, but it contains the wrong content (for the previous
story not the new one) so none of the dictionary words for the new story will
be highlighted (unless they share the same words).

This change uses requestAnimationFrame in all cases to wait for the DOM to
update before processing the content.