First implementation of story graph
This allows stories to be linked to others with some associated
information via the story_graph table.

A graph of the relationships is stored in the client and when
viewing a story it also provides links to the related stories.

This could also be used to create another story view option in
the find story page, where the story graph is listed, showing
all the connections between the stories.